requirements.txt
stable-baselines3>=2.0.0
shimmy>=2.0.0
tensorflow>=2.13.0
tensorboard>=2.13.0
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.25.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
gymnasium>=0.29.0
gym>=0.29.0
huggingface-hub>=0.16.0
plotly>=5.15.0
opencv-python>=4.8.0
pillow>=10.0.0
networkx>=3.1.0
